requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Core ML (CPU-only — no CUDA required, Python 3.13 compatible)
torch==2.7.1+cpu
torchvision==0.22.1+cpu

# Data & Training
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.9.0
seaborn>=0.13.2
Pillow>=10.4.0
albumentations>=1.4.6
opencv-python-headless>=4.10.0
tqdm>=4.66.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1
jinja2>=3.1.3

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.1
requests>=2.31.0
kaggle>=1.6.12
datasets>=2.18.0
